+++ web/src/main.js
@@ -9,8 +9,9 @@ import { library } from '@fortawesome/fontawesome-svg-core';
 import { faArrowRight, faArrowLeft } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 library.add(faArrowRight, faArrowLeft);
-Vue.component('font-awesome-icon', FontAwesomeIcon);
+Vue.component('FontAwesome', FontAwesomeIcon);
 
+// Prototype
 Vue.prototype.$d3 = d3;
 Vue.prototype.$moment = moment;