+++ web/vue.config.js
@@ -1,7 +1,7 @@
 
 module.exports = {
   publicPath: '/static/',
-  productionSourceMap: false,
+  productionSourceMap: true,
   devServer: {
     host: '0.0.0.0',
     port: 8082,